+++ src/main/java/io/konik/validation/AmountCalculator.java
@@ -48,8 +48,6 @@ import io.konik.zugferd.entity.trade.item.ItemTax;
 import io.konik.zugferd.entity.trade.item.SpecifiedAgreement;
 import io.konik.zugferd.entity.trade.item.SpecifiedMonetarySummation;
 import io.konik.zugferd.entity.trade.item.SpecifiedSettlement;
-import io.konik.zugferd.unece.codes.TaxCategory;
-import io.konik.zugferd.unece.codes.TaxCode;
 import io.konik.zugferd.unqualified.Amount;
 
 /**
@@ -92,7 +90,7 @@ public final class AmountCalculator {
 
     for (Item item : items) {
       log.info("==> {}:", ++itemsCounter);
-      log.info("Recalculating item: [{}]", 
+      log.info("Recalculating item: [{}]",
           item.getProduct() != null ? item.getProduct().getName() : "N/A");
 
       Amount lineTotal = new ItemLineTotalCalculator().apply(item);
@@ -553,68 +551,6 @@ public final class AmountCalculator {
     @Override
     public String toString() {
       return "TaxAggregator{" + "map=" + map + '}';
-    }
-
-    /**
-     * Helper key for {@link TaxAggregator}
-     */
-    static final class Key {
-      private final BigDecimal percentage;
-      private final TaxCode code;
-      private final TaxCategory category;
-
-      private Key(final Tax tax) {
-        this.percentage = tax.getPercentage();
-        this.category = tax.getCategory();
-        this.code = tax.getType();
-      }
-
-      public static Key create(final Tax tax) {
-        return new Key(tax);
-      }
-
-      public BigDecimal getPercentage() {
-        return percentage;
-      }
-
-      public TaxCode getCode() {
-        return code;
-      }
-
-      public TaxCategory getCategory() {
-        return category;
-      }
-
-      @Override
-      public boolean equals(Object o) {
-        if (this == o)
-          return true;
-        if (!(o instanceof Key))
-          return false;
-
-        Key key = (Key) o;
-
-        if (!percentage.equals(key.percentage))
-          return false;
-        if (code != key.code)
-          return false;
-        return category == key.category;
-
-      }
-
-      @Override
-      public int hashCode() {
-        int result = percentage.hashCode();
-        result = 31 * result + (code != null ? code.hashCode() : 0);
-        result = 31 * result + (category != null ? category.hashCode() : 0);
-        return result;
-      }
-
-      @Override
-      public String toString() {
-        return "Key{" + "percentage=" + percentage + ", code=" + code + ", category=" + category
-            + '}';
-      }
     }
   }
